JKI JSON Library: Add support for serializing and deserializing data stored in LabView Variants Attributes.
Variant Attributes can be used as key/value stores (hash tables) in LabVIEW.
Since attributes can be defined at runtime, they are also useful for constructing
nested data structures for which the data type is not known at compile time.
This commit adds support for serializing variant attributes to JSON strings, and
for deserializing JSON strings into LabView variant data with attributes. Arbitrary
JSON strings can be manipulated at run-time using a combination of LabView arrays
of variants and variant attributes.

Details:

A JSON string is either an object or an array comprising six kinds of values:
strings, numbers, objects, arrays true/false, and null. Most of the data handling
derives from the existing JKI JSON parser, the main exception being the use of
variant attributes instead of compile-time defined clusters for JSON objects.

Scalars
-------

Scalar types (strings, numbers, true/false) are represented as data of the
appropriate type cast to variants. Numeric types are represented by LabView
"extended" numeric type, which is at least 64 bits on all platforms.

Null
----
Null (also a scalar) is represented by a void (default) variant cast to a variant.
Compare with empty object below.

Arrays
------

Arrays are represented as arrays of variants cast to a (scalar) variant. While
LabView requires all elements of an array to have the same data type, JSON does
not: a string like [true, "Hamlet", 3.14159] is a legal JSON array. Thus the
library deserializes JSON arrays to an array of variants, which can then be cast as
necessary on a per-element basis. (LabView variants are data-type aware so you
can inspect the data type of each element inside a loop.)

Objects
-------

JSON objects are respresented by the library by a void (default) variant with
attributes. A void variant *without* any attributes defined is the empty object
but there is some ambiguity in how the library handles empty objects (see issues).

A void variant containing one or more attributes will be serialized as a JSON
object in which the JSON name is the attribute name and the JSON value the
attribute value. Objects can be nested by passing the appopriate input to Set
Variant Attribute function.

To extract data from variants with attributes, use the Get Variant Attribute
function in conjunection with Variant to Data. (Cast JSON arrays to an array
of variants if data type is not known as compile time). Nested objects can be
extracted with repeated application of the appropriate functions.

Issues:

1) LabView does not have a native null data type. This commit maintains the
behavior of the JKI JSON library in which JSON are represented in slightly
inconsistent ways, as follows.

On serialization, a default (void) variant generates a null in the output
JSON string. However, deserializing a null value returns a default (void)
variant cast to a variant (e.g. it is a nested structure.)

{"key":null}  returns a void variant cast to variant

This allows the deserializer to return a default (void) variant for an
empty object:

{"key": {} }  returns a void variant

From the point of view of usign variant attributes, the deserializer's
behavior is "correct", in that it returns a void variant with zero attributes,
but I did not want to change the behavior of the serializer without discussion.

2) I am making use of "Get Type Information.vi" found in LabView 2015 vi.lib. It
may be necessary to replace this function on earlier versions of LabView. (The
rest of the library uses a deprecated vi.lib function.)

3) Current the library is pretty difficult to use, in part because there
is no visibility into the deeper levels of the LabView variant structures on
the front panel. This will be addressed with "Generic JSON Data" class I'm
writing that provides object construction, searching etc.